dcbn cxz

+++ backends/sql/podatabase.class.php
@@ -1086,20 +1086,32 @@ class poDatabase extends database {
 			Podcasttable.Title AS album,
 			Podcasttable.Artist AS artist,
 			PodcastTracktable.Title AS title,
-			PodcastTracktable.Image AS image
+			PodcastTracktable.Image AS image,
+			PodcastTrackTable.PubDate AS pubdate,
+			PodcastTrackTable.Description AS comment,
+			PodcastTable.Image AS podimage
 			FROM PodcastTracktable
 			JOIN Podcasttable USING (PODindex)
 			WHERE PODTrackindex = ?",
 			intval($key)
 		);
 
+		$all_tracks = $this->sql_prepare_query(false, PDO::FETCH_COLUMN, 0, [],
+			"SELECT PubDate FROM PodcastTracktable WHERE PODindex = ? ORDER BY PubDate ASC",
+			intval($channel)
+		);
+
 		foreach ($result as $obj) {
 			$url = $obj->Link;
 			$filesize = $obj->FileSize;
 		}
+		$obj->podimage = str_replace('/small', '/asdownloaded/', $obj->podimage);
+		logger::log("PODCASTS", "     URL is",$url);
 		logger::log("PODCASTS", "  Artist is",$obj->artist);
 		logger::log("PODCASTS", "   Album is",$obj->album);
 		logger::log("PODCASTS", "   Title is",$obj->title);
+		logger::log("PODCASTS", "   Image is",$obj->image);
+		logger::log("PODCASTS", "PodImage is",$obj->podimage);
 		if ($url === null) {
 			logger::warn("PODCASTS", "  Failed to find URL for podcast",$channel);
 			header('HTTP/1.0 404 Not Found');
@@ -1109,8 +1121,9 @@ class poDatabase extends database {
 		// To make the progress bars look better in the GUI we attempt to read the actual filesize
 		$filesize = getRemoteFilesize($url, $filesize);
 		if (is_dir('prefs/podcasts/'.$channel.'/'.$key) || mkdir ('prefs/podcasts/'.$channel.'/'.$key, 0755, true)) {
-			$filename = basename($url);
-			$filename = preg_replace('/\?.*$/','',$filename);
+			$filename = basename(rawurldecode($url));
+			$filename = format_for_disc($filename);
+			// $filename = preg_replace('/\?.*$/','',$filename);
 
 			$xml = '<?xml version="1.0" encoding="utf-8"?><download><filename>';
 			$xml = $xml . 'prefs/podcasts/'.$channel.'/'.$key.'/'.$filename;
@@ -1143,25 +1156,35 @@ class poDatabase extends database {
 				if (array_key_exists('tags', $getID3->info)) {
 					$current_tags = $getID3->info['tags'];
 					if (is_array($current_tags) && array_key_exists('id3v2', $current_tags)) {
-						logger::debug('PODCASTS', 'Using Current ID3v2 Tags');
+						logger::trace('PODCASTS', 'Using Current ID3v2 Tags');
 						$tags = $current_tags['id3v2'];
 					} else if (is_array($current_tags) && array_key_exists('id3v1', $current_tags)) {
-						logger::debug('PODCASTS', 'Using Current ID3v1 Tags');
+						logger::trace('PODCASTS', 'Using Current ID3v1 Tags');
 						$tags = $current_tags['id3v1'];
 					}
 				}
 
+				$track_number = 0;
 				if (array_key_exists('track_number', $tags) && is_array($tags['track_number'])) {
 					$track_number = $tags['track_number'][0];
-					logger::debug('PODCASTS', 'Using track number from tags',$track_number);
-				} else {
+					logger::trace('PODCASTS', 'Using track number from tags',$track_number);
+				}
+				if ($track_number == 0) {
 					$track_number = format_tracknum($obj->title);
-					logger::debug('PODCASTS', 'Using track number from title',$track_number);
 					if ($track_number > 0)
-						$tags['track_number'] = array($track_number);
+						logger::trace('PODCASTS', 'Using track number from title',$track_number);
+				}
+				if ($track_number == 0) {
+					$track_number = (array_search($obj->pubdate, $all_tracks)) + 1;
+					if ($track_number > 0)
+						logger::trace('PODCASTS', 'Using track number from published order',$track_number);
 				}
 
+				if (is_numeric($track_number) && $track_number > 0) {
+					$tags['track_number'] = [$track_number];
+				}
 				$track_image_file = null;
+				$has_trackimage = false;
 				if ($obj->image) {
 					logger::trace('PODCASTS', 'Downloading Image',$obj->image);
 					$imgd = new url_downloader(['url' => $obj->image]);
@@ -1169,10 +1192,14 @@ class poDatabase extends database {
 					$track_image_file = 'prefs/podcasts/'.$channel.'/'.$key.'/cover.'.$ext;
 					if ($imgd->get_data_to_file($track_image_file, true)) {
 						logger::log('PODCASTS', 'Downloaded Track Image to Embed into file');
+						$has_trackimage = true;
 					} else {
 						logger::warn('PODCASTS', 'Failed to download track image');
 						$track_image_file = null;
 					}
+				} else if (file_exists($obj->podimage)) {
+					logger::trace('PODCASTS', 'Embedding Podcast image', $obj->podimage);
+					$track_image_file = $obj->podimage;
 				}
 
 				if ($obj->WriteTags != 0) {
@@ -1181,6 +1208,7 @@ class poDatabase extends database {
 					$tags['albumartist'] = array($obj->artist);
 					$tags['album'] = array($obj->album);
 					$tags['title'] = array($obj->title);
+					$tags['comment'] = array($obj->comment);
 				}
 
 				if ($track_image_file) {
@@ -1198,13 +1226,13 @@ class poDatabase extends database {
 					}
 				}
 
-				if ($obj->WriteTags != 0 || $track_image_file != null) {
+				if ($obj->WriteTags != 0 || $has_trackimage) {
 					logger::log('PODCASTS', 'Writing ID3 tags to',$download_file);
 					getid3_lib::IncludeDependency(GETID3_INCLUDEPATH.'write.php', __FILE__, true);
 
 					$tagwriter = new getid3_writetags();
 					$tagwriter->filename       = $download_file;
-					$tagwriter->tagformats     = array('id3v2.3');
+					$tagwriter->tagformats = array('id3v2.3');
 					$tagwriter->overwrite_tags = true;
 					$tagwriter->tag_encoding   = 'UTF-8';
 					$tagwriter->remove_other_tags = true;
@@ -1220,7 +1248,7 @@ class poDatabase extends database {
 				} else {
 					logger::log('PODCASTS', 'No Tags to Write');
 				}
-				if ($track_image_file)
+				if ($track_image_file && $has_trackimage)
 					unlink($track_image_file);
 
 				if ($track_number > 0 && is_numeric($track_number)) {
